requirements.txt
flask
chromadb
google-genai
beautifulsoup4
python-dotenv
requests
sentence_transformers